requirements.txt
openenv-core>=0.2.2
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
numpy>=1.24.0
huggingface-hub>=0.22.0
python-multipart>=0.0.6
httpx>=0.24.0
